Initial migration and first "official" release.
NOTICE: This repository is meant to be a basic starting point for
websites which are intended to require little to no maintenance.
Carefully contemplate any potential changes and consider if they are
tweaks for the sake of gold plating or if they will have a serious
impact on end users (consider the first 164).